refactor: split settingsSlice file to canvasSlice and selectedImageSlice
